refactor(bspline): split the basis kernels out of their Layer 2 module
`_bspline_basis_core.py` held both the Numba kernels and the Layer 2 entry
points that ask `_basis_backend` which kernel to call. That closed a cycle, so
the catalogue could not import the kernels at module scope the way every other
catalogue in the package does, and deferred both imports into its two
functions with a comment explaining why.

`_bspline_basis_kernels.py` now holds the kernels and nothing else, the way
`_bspline_extraction_core.py` does for the extraction port. Nothing there
imports the catalogue, so the catalogue imports it directly and the two
deferred imports are gone along with the paragraph that justified them.

The kernels moved line for line; what changed is which file they are in, and
`_warmup_numba_functions` moved with them because it exists to compile them.

The downstream consumer was re-censused first, since these are private symbols
and `CLAUDE.md` records that its largest exposure is `pantr.bspline`. Three
sibling checkouts, 58 distinct import lines, zero private symbols: no
underscore-prefixed module path and no underscore-prefixed name in any import
list. The only submodule it reaches by path is public.
